moved imageio classes to a subpackage
